tratamento de erros na Delete

+++ src/main/java/fcss_dev/gas_station/applitation/service/TipoCombustivelService.java
@@ -50,6 +50,9 @@ public class TipoCombustivelService {
 
     // Delete
     public void deletar(Long id){
+        if (!repository.existsById(id)) {
+            throw new NenhumRegistroEncontradoException("Não foi possível excluir: ID " + id + " não existe");
+        }
         repository.deleteById(id);
     }
 }
